apps/backend/pyproject.toml
[project]
name = "pcqvp-backend"
version = "0.1.0"
description = "PCQVP Platform - Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.0",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]>=2.0.48",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "python-dotenv",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "email-validator",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "slowapi>=0.1.9",
    "fastapi-mail>=1.4.0",
    "itsdangerous>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
